+++ src/core/shards/channels.cpp
@@ -1,80 +1,10 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /* Copyright © 2020 Fragcolor Pte. Ltd. */
 
-#include "shared.hpp"
-#include <atomic>
-#include <boost/lockfree/queue.hpp>
-#include <boost/lockfree/stack.hpp>
-#include <mutex>
-#include <variant>
+#include "channels.hpp"
 
 namespace shards {
 namespace channels {
-
-struct ChannelShared {
-  SHTypeInfo type;
-  std::atomic_bool closed;
-};
-
-struct DummyChannel : public ChannelShared {};
-
-struct MPMCChannel : public ChannelShared {
-  MPMCChannel(bool noCopy) : ChannelShared(), _noCopy(noCopy) {}
-
-  // no real cleanups happens in Produce/Consume to keep things simple
-  // and without locks
-  ~MPMCChannel() {
-    if (!_noCopy) {
-      SHVar tmp{};
-      while (data.pop(tmp)) {
-        destroyVar(tmp);
-      }
-      while (recycle.pop(tmp)) {
-        destroyVar(tmp);
-      }
-    }
-  }
-
-  // A single source to steal data from
-  boost::lockfree::queue<SHVar> data{16};
-  boost::lockfree::stack<SHVar> recycle{16};
-
-private:
-  bool _noCopy;
-};
-
-struct Broadcast;
-class BroadcastChannel : public ChannelShared {
-public:
-  BroadcastChannel(bool noCopy) : ChannelShared(), _noCopy(noCopy) {}
-
-  MPMCChannel &subscribe() {
-    // we automatically cleanup based on the closed flag of the inner channel
-    std::scoped_lock<std::mutex> lock(subMutex);
-    return subscribers.emplace_back(_noCopy);
-  }
-
-protected:
-  friend struct Broadcast;
-  std::mutex subMutex;
-  std::list<MPMCChannel> subscribers;
-  bool _noCopy = false;
-};
-
-using Channel = std::variant<DummyChannel, MPMCChannel, BroadcastChannel>;
-
-class Globals {
-private:
-  static inline std::mutex ChannelsMutex;
-  static inline std::unordered_map<std::string, Channel> Channels;
-
-public:
-  static Channel &get(const std::string &name) {
-    std::scoped_lock<std::mutex> lock(ChannelsMutex);
-    return Channels[name];
-  }
-};
-
 struct Base {
   Channel *_channel = nullptr;
   std::string _name;
